requirements.txt
PySide6
PySide6-Addons
python-vlc
requests
python-dotenv